requirements.txt
Pillow>=10.0.0
colorama>=0.4.6
imageio>=2.22.0
numpy>=1.24.0
pandas>=2.0.0
gymnasium>=0.29.0
stable-baselines3>=2.0.0
simglucose
torch>=2.0.0
scipy>=1.10.0
optuna>=3.0.0
matplotlib

